opentelemetry-first/pyproject.toml
[tool.poetry]
name = "opentelemetry-first"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
opentelemetry-api = "^1.29.0"
opentelemetry-sdk = "^1.29.0"
opentelemetry-exporter-prometheus = "^0.50b0"
opentelemetry-semantic-conventions = "^0.50b0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
